# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: SwipeSavvy Backend API
  description: |
    Multi-service backend API for SwipeSavvy Mobile Wallet — a fintech platform providing
    card issuance (FIS), payments (Authorize.Net), KYC, rewards, real-time chat, and admin management.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `POST /api/v1/auth/login` + `POST /api/v1/auth/verify-login-otp`.

    | Token | Lifetime | Purpose |
    |-------|----------|---------|
    | Access Token | 24 hours | Authorize API requests |
    | Refresh Token | 30 days | Obtain new access tokens |

    ## Environments
    Use the server selector above to switch between Production, Staging, and Development.
  version: 1.0.0
  contact:
    name: SwipeSavvy Engineering
    email: engineering@swipesavvy.com
  license:
    name: Proprietary
servers:
  - url: http://54.224.8.14:8000
    description: Production (AWS EC2)
  - url: https://api.swipesavvy.com
    description: Production (Domain)
  - url: http://localhost:8000
    description: Development (Local)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login flow
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token from admin login
    WebhookSignature:
      type: apiKey
      in: header
      name: X-FIS-Signature
      description: FIS webhook signature verification
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        date_of_birth:
          type: string
          format: date
        email_verified:
          type: boolean
        phone_verified:
          type: boolean
        kyc_tier:
          type: string
          enum: [none, tier1, tier2, tier3]
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 86400
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login with email and password (triggers OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  otp_required:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Invalid credentials
        '423':
          description: Account locked
...
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset
        '400':
          description: Invalid token
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid refresh token
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
...
  /api/v1/admin/auth/login:
    post:
      tags: [Admin Auth]
      operationId: adminLogin
      summary: Admin login (rate limited 5/min/IP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Admin session
        '401':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded
  /api/v1/admin/auth/refresh:
    post:
      tags: [Admin Auth]
      operationId: adminRefresh
      summary: Refresh admin token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: New token
  /api/v1/admin/auth/logout:
    post:
      tags: [Admin Auth]
      operationId: adminLogout
      summary: Admin logout
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Logged out
  /api/v1/admin/auth/me:
    get:
      tags: [Admin Auth]
      operationId: adminGetProfile
      summary: Get admin profile
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Admin profile
  /api/v1/admin/auth/setup-admin:
    post:
      tags: [Admin Auth]
      operationId: setupAdmin
      summary: Initial admin setup (requires setup key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [setup_key, email, password, full_name]
              properties:
                setup_key:
                  type: string
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
      responses:
        '200':
          description: Admin created
  /api/v1/admin/auth/reset-password:
    post:
      tags: [Admin Auth]
      operationId: adminResetPassword
      summary: Reset admin password (requires setup key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [setup_key, email, new_password]
              properties:
                setup_key:
                  type: string
                email:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password reset
...